Conflicts:
	src/stan/meta/traits.hpp
	src/stan/prob/distributions/multivariate/continuous/multi_normal.hpp
	src/stan/prob/distributions/multivariate/continuous/multi_normal_cholesky.hpp
	src/stan/prob/distributions/multivariate/continuous/multi_normal_prec.hpp
	src/stan/prob/distributions/multivariate/continuous/multi_student_t.hpp

+++ src/test/unit/mcmc/chains_test.cpp
@@ -156,8 +156,8 @@ TEST_F(McmcChains, add_adapter) {
     }
   }
 
-  EXPECT_EQ(blocker1.samples.rows(), samples.size());
-  EXPECT_EQ(blocker1.samples.cols(), samples[0].size());
+  EXPECT_EQ(blocker1.samples.rows(), static_cast<int>(samples.size()));
+  EXPECT_EQ(blocker1.samples.cols(), static_cast<int>(samples[0].size()));
 
   EXPECT_NO_THROW(chains.add(samples))
     << "adding multiple samples, adds new chain";
